refactor(service): 纯移动拆分 openai_gateway_service.go（4872→1095行）
